Folder structure is initialized
