+++ src/service/auth.js
@@ -1,5 +1,5 @@
 import router from '@/router'
-import store from '.'
+import store from '@/store/'
 export default {
   namespaced: true,
   state: {},
